fix: move provider resolution to remote-server
- keep engine defaults untouched

- build provider factory in remote-server
